BCC from ADO to start the process of adding BC support
Target goal for today was to get `src-cli` to recognize Perforce as a supported code host.

At the beginning of the day, using `src-cli` as per [the documentation](https://docs.sourcegraph.com/batch_changes/quickstart)
resulted in the error `Some repositories are hosted on unsupported code hosts and will be skipped. Use the -allow-unsupported flag to avoid skipping them.`

Tracing that error message through to GraphQL showed that all that was neede to get past that error was adding `extsvc.TypePerforce: {},` to the `SupportedExternalServices` map in `enterprise/internal/batches/types/util.go`.

Along the way, Perforce BC code was blind carbon-copied from the Azure Dev Ops BC implementation, which is the most recent prior BC implementation.

Now that the initial "unsupported code host" error has been resolved, the next error is `fetching repo: fetching ZIP archive: unable to fetch archive (HTTP 500 from https://sourcegraph.test:3443/...`

Not sure why it's erroring - visiting that URL from the browser results in a text/plain document withe the names of the files in the repo. That's no an HTTP 500 error.

Ah well, that's for next time.

The mid-range goal is to get to the point of being able to execute a batch change process in a worker, which should result in a changelist being shelved in the Perforce server.