internal/database: Use perforce_changelist_id to get the latest row of a repo
Because perforce uses incremental numbers, this guarantees that the
highest changelist ID of a repo will be the latest git commit.

We cannot use ID because our insertion starts from the latest to the
oldest commit. While this may be helpful for the first time a repo is
imported, subsequent imports of new commits will break this expectation
immediately is the latest ID will be the oldest commit in the newly
imported set of commits.